Streaming service is working now
To start or test the streaming you have to execute the command below on the js folder. (really important)

node server.js

After that, execute the stream.html file for broadcast your webcam, and when you call the "wacheando-kmera" div you will be able to stream the video.

That is how it works.